Added JLabel wrapped in a JPanel that shows the current coordinates.
There seems to be a problem where the coordinate for centerY is the
negative of what it should be (i.e. -0.7 instead of 0.7). It may have to
do with the fact that the graphics treats the upper left corner as the
origin while the calculations treat the bottom right as the origin.